finish

+++ app/src/main/java/com/phamnhantucode/cryptotracker/crypto/data/networking/dto/CoinPriceDto.kt
@@ -4,6 +4,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class CoinPriceDto(
-    val priceUsd: Double,
+    val priceUsd: Float,
     val time: Long
 )
