+++ app/src/main/java/com/phamnhantucode/cryptotracker/crypto/domain/CoinPrice.kt
@@ -3,6 +3,6 @@ package com.phamnhantucode.cryptotracker.crypto.domain
 import java.time.ZonedDateTime
 
 data class CoinPrice(
-    val priceUsd : Double,
+    val priceUsd : Float,
     val time: ZonedDateTime
 )
